+++ src/main/java/org/spout/vanilla/world/generator/normal/biome/TaigaBiome.java
@@ -26,13 +26,11 @@
  */
 package org.spout.vanilla.world.generator.normal.biome;
 
-
 import net.royawesome.jlibnoise.module.modifier.ScalePoint;
 
 import org.spout.vanilla.configuration.VanillaConfiguration;
 
 public class TaigaBiome extends VanillaNormalBiome {
-	
 	private final static ScalePoint NOISE = new ScalePoint();
 
 	static {
@@ -41,13 +39,13 @@ public class TaigaBiome extends VanillaNormalBiome {
 		NOISE.setyScale(VanillaConfiguration.BIOMES.TAIGA_Y_SCALE.getDouble());
 		NOISE.setzScale(VanillaConfiguration.BIOMES.TAIGA_Z_SCALE.getDouble());
 	}
-	
+
 	public TaigaBiome(int biomeId) {
 		super(biomeId, NOISE);
 		this.minDensityTerrainHeight = VanillaConfiguration.BIOMES.TAIGA_MIN_DENSITY_TERRAIN_HEIGHT.getByte();
 		this.maxDensityTerrainHeight = VanillaConfiguration.BIOMES.TAIGA_MAX_DENSITY_TERRAIN_HEIGHT.getByte();
-		this.upperHeightMapScale = VanillaConfiguration.BIOMES.TAIGA_UPPER_HEIGHT_MAP_SCALE.getInt();
-		this.bottomHeightMapScale = VanillaConfiguration.BIOMES.TAIGA_BOTTOM_HEIGHT_MAP_SCALE.getInt();
+		this.upperHeightMapScale = VanillaConfiguration.BIOMES.TAIGA_UPPER_HEIGHT_MAP_SCALE.getFloat();
+		this.bottomHeightMapScale = VanillaConfiguration.BIOMES.TAIGA_BOTTOM_HEIGHT_MAP_SCALE.getFloat();
 	}
 
 	@Override
